change page directory
use await instead

test api using pipeline address
